keylogger/pyproject.toml
[project]
name = "seg-comp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
	"asyncio>=3.4.3",
	"getmac>=0.9.5",
	"pynput>=1.8.1",
	"python-dotenv>=1.1.0",
	"requests>=2.32.3",
]

[project.scripts]
start = "main:__main__"

[dependency-groups]
dev = ["pyright>=1.1.398", "ruff>=0.11.4"]
